move updateCartItemCount() to utls, and import to other js files.
