+++ src/js/productDetails.mjs
@@ -1,13 +1,14 @@
 import {
   qs,
-  setLocalStorage
+  setLocalStorage,
+  getParam,
+  updateCartItemCount,
 } from "./utils.mjs";
+
 import {
   findProductById
 } from "./productData.mjs";
-import {
-  getParam
-} from "./utils.mjs";
+
 
 export default async function productDetails(productId, selector) {
   const product = await findProductById(productId);
@@ -64,23 +65,6 @@ function addToCart(id, product) {
 
 }
 
-// update cart item count in header
-export function updateCartItemCount() {
-  const cartItems = Object.keys(localStorage);
-  const cartItemCount = cartItems.length;
-  const cartItemCountElement = document.getElementById("cartItemCount");
-
-  if (cartItemCountElement !== null) {
-    if (cartItemCount === 0) {
-      // If there aren't items in the cart, hide the cartItemCount element
-      cartItemCountElement.style.display = "none";
-    } else {
-      // If there are items in the cart, show the cartItemCount element
-      cartItemCountElement.style.display = "block";
-      cartItemCountElement.textContent = cartItemCount;
-    }
-  }
-}
 
 // Call function so it runs on page load and updates cart count
 updateCartItemCount();
